[UPDATE] Removed all business logic in reducers to specific helper functions, converted a few more components to functions. Styling data to be ported next.
