+++ website/components/layout.tsx
@@ -3,15 +3,14 @@ import {} from "@material-ui/icons";
 //import Navbar from "./Navbar";
 import Newappbar from "./newAppbar";
 
-class Layout extends React.Component {
-  render() {
+function Layout(props) {
+  
     return (
       <React.Fragment>
         <Newappbar />
-        {this.props.children}
+        {props.children}
       </React.Fragment>
     );
   }
-}
 
 export default Layout;
