+++ website/store/actionTypes/contestActionType.tsx
@@ -4,11 +4,6 @@ export const CONTEST_REQUEST_PENDING = "CONTEST_REQUEST_PENDING";
 export const CONTEST_DATA_SUCCESS = "CONTEST_DATA_SUCCESS";
 export const CONTEST_DATA_FAILURE = "CONTEST_DATA_FAILURE";
 
-//Leaderboard fetching actions
-export const LEADERBOARD_QUES_REQUEST_PENDING = "LEADERBOARD_QUES_REQUEST_PENDING";
-export const LEADERBOARD_QUES_DATA_FAILURE = "LEADERBOARD_QUES_DATA_FAILURE";
-export const LEADERBOARD_QUES_DATA_SUCCESS = "LEADERBOARD_QUES_DATA_SUCCESS";
-
 export const LEADERBOARD_REQUEST_PENDING = "LEADERBOARD_REQUEST_PENDING";
 export const LEADERBOARD_DATA_FAILURE = "LEADERBOARD_DATA_FAILURE";
 export const LEADERBOARD_DATA_SUCCESS = "LEADERBOARD_DATA_SUCCESS";
