+++ website/store/actions/leaderboardAction.tsx
@@ -1,7 +1,4 @@
 import {
-    LEADERBOARD_QUES_REQUEST_PENDING,
-    LEADERBOARD_QUES_DATA_SUCCESS,
-    LEADERBOARD_QUES_DATA_FAILURE,
     LEADERBOARD_REQUEST_PENDING,
     LEADERBOARD_DATA_SUCCESS,
     LEADERBOARD_DATA_FAILURE,
@@ -11,41 +8,25 @@ import { contestService } from "../../services/contestService"
 
 export function getLearderboardDataQues() {
     return function(dispatch) {
-        dispatch({ type: LEADERBOARD_QUES_REQUEST_PENDING });
+        dispatch({ type: LEADERBOARD_REQUEST_PENDING });
         return contestService.getLeaderboardDataQues()
             .then((result) => {
-                // if (result.data.detail) {
-                //     dispatch({ type: LEADERBOARD_QUES_DATA_FAILURE, payload: result.data });
-                // } else {
-                    dispatch({ type: LEADERBOARD_QUES_DATA_SUCCESS, payload: result })
-                // }
-            })
-            .catch(() => {
-                // console.log(err)
-                dispatch({
-                    type: LEADERBOARD_QUES_DATA_FAILURE,
-                    payload: { detail: "Something wrong while fetching question data for leaderboard" },
-                });
-            });
-    };
-}
-
-export function getLearderboardData() {
-    return function(dispatch) {
-        dispatch({ type: LEADERBOARD_REQUEST_PENDING });
-        return contestService.getLeaderboardData()
-            .then((result) => {
-                // if (result.data.detail) {
-                //     dispatch({ type: LEADERBOARD_DATA_FAILURE, payload2: result.data });
-                // } else {
-                    dispatch({ type: LEADERBOARD_DATA_SUCCESS, payload: result.data })
-                // }
-            })
-            .catch(() => {
-                // console.log(err)
+                return contestService.getLeaderboardData().then((result2) => {
+                    var arr=[]
+                    arr.push(result2.data,result);
+                    dispatch({ type: LEADERBOARD_DATA_SUCCESS, payload: arr })
+            }).catch(() => {
                 dispatch({
                     type: LEADERBOARD_DATA_FAILURE,
                     payload: { detail: "Something wrong while fetching leaderboard data" },
+                });
+            });
+             // dispatch({ type: LEADERBOARD_QUES_DATA_SUCCESS, payload: result })
+            })
+            .catch(() => {
+                dispatch({
+                    type: LEADERBOARD_DATA_FAILURE,
+                    payload: { detail: "Something wrong while fetching question data for leaderboard" },
                 });
             });
     };
