+++ website/store/reducers/personalSubmissionsReducer.tsx
@@ -4,6 +4,8 @@ import {
     PERSONAL_SUBMISSIONS_DATA_FAILURE,
 } from "../actionTypes/contestActionType";
 
+import {middleware} from "../../helper/middleware";
+
 const initialState = {
     personal_submissions: [],
     loaded: false,
@@ -24,53 +26,9 @@ const initialState = {
 
         case PERSONAL_SUBMISSIONS_DATA_SUCCESS:
             {
-                var arr = [];
-                action.payload.map((r) => {
-                    var stat = "";
-                    var time = "";
-                    var mem = "";
-                    const cases = JSON.parse(r.status);
-                    cases.map((testcase) => {
-                      if (testcase.code == 1) {
-                        stat = "Compilation Error";
-                        time = "NA";
-                        mem = "NA";
-                      } else {
-                        if (testcase.status.run_status == "AC") {
-                          if (stat == "") {
-                            stat = "AC";
-                            time = testcase.status.cpu_time + " sec";
-                            mem = testcase.status.memory_taken + " kb ";
-                          }
-                        } else {
-                          stat = testcase.status.run_status;
-                          time = testcase.status.cpu_time + " sec";
-                          mem = testcase.status.memory_taken + " kb";
-                        }
-                      }
-                    });
-          
-                    var runtimestats = {
-                      source: r.code,
-                      testcases: cases,
-                      lang: r.lang,
-                    };
-          
-                    var payload = {
-                      user: r.name,
-                      problem: r.question_name,
-                      status: stat,
-                      time: time,
-                      memory: mem,
-                      code: runtimestats,
-                    };
-          
-                    arr.push(payload);
-                  });
-
-                  console.log(arr);
+                
                 return Object.assign({}, state, {
-                    personal_submissions: arr,
+                    personal_submissions: middleware.processPS(action.payload),
                     loaded: true,
                 })
             }
